requirements.txt
Pillow==9.1.1
CairoSVG==2.5.2
tqdm==4.64.1
